Port the function classes
